sped up the game just a tad. Fixed graphic artifacts with font sizes. Removed cached font sizes because it wasn't necessary.

+++ iBrogue_iPad/PlatformCode/Viewport.h
@@ -29,9 +29,9 @@
 #define FONT_SIZE	13
 
 typedef struct {
-    short red;
-    short green;
-    short blue;
+    float red;
+    float green;
+    float blue;
 } SHColor;
 
 @interface Viewport : ACMagnifyingView
